+++ camdecmps/procedures/load_temp_daily_test_errors.sql
@@ -19,7 +19,7 @@ BEGIN
       DAILY_TEST_SUM_ID character varying(45) NOT NULL,
       MON_PLAN_ID character varying(45) NOT NULL, 
       MON_LOC_ID character varying(45) NOT NULL, 
-      RPT_PERIOD_ID bigint NOT NULL, 
+      RPT_PERIOD_ID numeric(38,0) NOT NULL,
       COMPONENT_ID character varying(45), 
       COMPONENT_IDENTIFIER character varying(3),
       COMPONENT_TYPE_CD character varying(7),
