+++ camdecmpscalc/tables/system_op_supp_data.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS camdecmpscalc.system_op_supp_data
 (
     pk character varying(45) COLLATE pg_catalog."default" NOT NULL DEFAULT uuid_generate_v4(),
     mon_sys_id character varying(45) COLLATE pg_catalog."default" NOT NULL,
-    rpt_period_id bigint NOT NULL,
+    rpt_period_id numeric(38,0) NOT NULL,
     op_supp_data_type_cd character varying(7) COLLATE pg_catalog."default" NOT NULL,
     days numeric(38,0) NOT NULL,
     hours numeric(38,0) NOT NULL,
