+++ camdecmpswks/tables/monitor_plan.sql
@@ -13,8 +13,8 @@ CREATE TABLE IF NOT EXISTS camdecmpswks.monitor_plan
     submission_id bigint,
     submission_availability_cd character varying(7) COLLATE pg_catalog."default" DEFAULT 'GRANTED'::character varying,
     pending_status_cd character varying(7) COLLATE pg_catalog."default",
-    begin_rpt_period_id bigint NOT NULL,
-    end_rpt_period_id bigint,
+    begin_rpt_period_id numeric(38,0) NOT NULL,
+    end_rpt_period_id numeric(38,0),
     last_evaluated_date timestamp without time zone,
     eval_status_cd character varying(7) COLLATE pg_catalog."default" NOT NULL DEFAULT 'EVAL'::character varying
 );
